Fix numerous bugs in piano implementation
Most of these bugs resulted in vector out-of-range exceptions. A bug in ShiftView caused the method to shift the view by semitones rather than white keys. A bug in SetKeyBinds() caused out-of-range exceptions because it did not compare its index counter against the size of the array it was iterating through.