Fix bug that caused offset sound files
Bug resulted from a inaccurate assumption in MusicNote class. In music, every octave's first note is a C and every piano starts with A0. MusicNote assumed that every octave's first note was an A. This caused Piano to build keys like C0, D0, which don't exist on the standard piano. This was fixed by resetting the wholetone indices to start on C. Important: The first semitone index on a standard piano is now 9 (before this change, the first semitone was 0).

Tests have been updated to reflect this change.