Fix major bug that prevented note from double-playing
Before, note voices would be played only if the note was currently enabled and the note voice would be disabled if the note was enabled. On StopNote(), the voice would actually stop with a 1 second delay to force the sound the resonate. During this duration, the voice would still be enabled, preventing the voice from being played again during the 1 second duration. An unsuccessful fix would be to remove the isEnabled() condition entirely. Doing this would cause the voice would rapidly repeat while a key was held down. The bug was resolved by creating a custom boolean (is_playing_) that is toggled while the note is held down. When the note is released, the boolean is set to false even though the sound is still resonating. Then, on a second click, the the toggle is simply flipped to true and plays a new sound over the resonating sound.